Fix Texture
git-svn-id: https://free1.projectlocker.com/mortennobel/RenderE/svn/trunk@53 75417771-6e0e-4020-8c6c-34838eb40c82

+++ src/render_e/textures/TextureBase.cpp
@@ -10,7 +10,7 @@
 
 namespace render_e {
 
-TextureBase::TextureBase(unsigned short textureType)
+TextureBase::TextureBase(int textureType)
 :textureType(textureType), mipmapping(true), textureId(0){
 }
